Authorization

+++ controllers/note.js
@@ -66,6 +66,10 @@ exports.getNote = (req, res, next) => {
 exports.deleteNote = (req, res, next) => {
     const { id } = req.params;
     Note.findById(id).then((note) => {
+        // in middleware, req.userId =
+        if (note.creater.toString() !== req.userId) {
+            return res.status(401).json("Auth Failed")
+        }
         if (note.cover_image) {
             unlink(note.cover_image);
         }
@@ -85,6 +89,9 @@ exports.deleteNote = (req, res, next) => {
 exports.getOldNote = (req, res, next) => {
     const { id } = req.params;
     Note.findById(id).then(note => {
+        if (note.creater.toString() !== req.userId) {
+            return res.status(401).json("Auth Failed")
+        }
         res.status(200).json(note);
     }).catch(err => {
         console.log(err)
@@ -100,6 +107,9 @@ exports.updateNote = (req, res, next) => {
     const cover_image = req.file;
 
     Note.findById(note_id).then(note => {
+        if (note.creater.toString() !== req.userId) {
+            return res.status(401).json("Auth Failed")
+        }
         note.title = title;
         note.content = content;
         if (cover_image) {
